refactor(OrganizationSelector):extract to own file
